requirements.txt
flask
flask_cors
requests
requests[security]
beautifulsoup4
pygooglevoice

